feat: add test of useConnectedMetaMask
