+++ package.json
@@ -57,7 +57,7 @@
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-testing-library": "^3.10.1",
-    "eth-testing": "^1.1.0-alpha.4",
+    "eth-testing": "^1.1.0-alpha.15",
     "husky": "^4.3.7",
     "jest": "^26.6.3",
     "lint-staged": "^10.5.3",
